Improve SQLiteConnectionHelper for multiple selection.
I need to explain the changes a bit: In the near future I want that every change on one or multiple items should be revertable. In the current state the multiple selection just executes the relevant ICommand implementation for the item. But because the item will also perform each change (e.g. marking it as read) as seperate revertable action, the `_transactionPoint` would be overwritten every time and only the latest change could be reverted. This is definitely not wanted, therefore the code checks if the execution of offline tasks were already blocked and if that's the case, the `_transactionPoint` won't be overwritten. Furthermore it won't start a new `Task.Delay()` which then would result to multiple task cancellations instead of just one.