Execute task inside the AddAsync method.
This way has multiple advantages. First of all, it's a far cleaner and easier to control approach than subscribing to the event. Second, it should (never tested it) return the valid result of the task status, so listening to that should five better results. Third, it allows flexibility for the future.